Module northstar::npk::npk

NPK file format

Structs

Hashes

NPK Hashes

Meta

NPK archive comment

Npk

Northstar package

SquashfsOpts

Squashfs Options

Enums

CompressionAlgorithm

Squashfs compression algorithm

Error

NPK loading error

Constants

VERSION

Manifest version supported by the runtime

Functions

generate_key

Generate a keypair suitable for signing and verifying NPKs

open

Open a Zip file

pack

Create an NPK for the northstar runtime. sextant collects the artifacts in a given container directory, creates and signs the necessary metadata and packs the results into a zipped NPK file.

pack_with

Create an NPK with special squashfs options sextant collects the artifacts in a given container directory, creates and signs the necessary metadata and packs the results into a zipped NPK file.

unpack

Extract the npk content to out