Module northstar::seccomp

Support for seccomp syscall filtering

Modules

profiles

Predefined seccomp profiles

Structs

AllowList

Read-only list of allowed syscalls. Methods do not cause memory allocations on the heap.

Seccomp

Seccomp configuration

Selinux

SELinux configuration

SyscallArgRule

Syscall argument rule

Enums

Profile

Predefined seccomp profile

SyscallRule

Syscall rule

Functions

seccomp_filter

Construct a allowlist syscall filter that is applied post clone.